moved order related controller functions into controller/orders.js
